Revert "Updated jar2lib dependency to 1.1.2-SNAPSHOT"
This reverts commit ceea26da54a03b8af7b3fcf0bf6217ba96336588.

We *do* want to depend on release versions *only*, *always*.

Otherwise builds are not reproducible. To couple cppwrap-maven-plugin
with a SNAPSHOT version of jar2lib for development of both components
simultaneously, modify the POM locally *only* and do not commit the
changes. If the changes must be committed (e.g., to collaborate on a
work in progress with other developers or across machines), commit the
POM change as a WIP on a topic branch (*never* to master) and make sure
the WIP has been purged from the history before merging to master.